% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cateDiffs.R
\name{cateDiffs}
\alias{cateDiffs}
\title{cateDiffs
Estimates CATE using causal forests with data from multiple trials
and carries out multiple treatment effect heterogeneity diagnostics by subgroup.}
\usage{
cateDiffs(
  dfList,
  outCol,
  txCol,
  covList,
  blpredList = NULL,
  combine = FALSE,
  ci = 0.95,
  nTrees = 10000,
  seedN = 7203
)
}
\arguments{
\item{dfList}{list; a list of dataframes, one per trial}

\item{outCol}{character; name of outcome column}

\item{txCol}{character; name of treatment status column}

\item{covList}{vector; column names of covariates to estimate forest}

\item{blpredList}{vector; number of units to be initially randomly selected}

\item{combine}{logical; combine data from multiple studies}

\item{ci}{numeric; confidence interval \%}

\item{nTrees}{numeric; column names of unit level variables  to sample units on}

\item{seedN}{numeric; seed number to be used for sampling. If NA, calls set.seed(); default = NA}
}
\value{
list with: 1) , 2),  3)
}
\description{
cateDiffs
Estimates CATE using causal forests with data from multiple trials
and carries out multiple treatment effect heterogeneity diagnostics by subgroup.
}
