% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBLPfig.R
\name{getBLPfig}
\alias{getBLPfig}
\title{Takes the best linear projection* of the CATE as a function
of of user-provided covariates and visually displays resulting coefficients.
\itemize{
\item Let tau(Xi) = E(Y(1) - Y(0) | X = Xi) be the CATE, and Ai be a vector of user-provided covariates.
This function provides a (doubly robust) fit to the linear model tau(Xi) ~ beta_0 + Ai * beta."
}}
\usage{
getBLPfig(res, combine)
}
\arguments{
\item{res}{list;}

\item{combine}{logical; combine data from multiple studies}
}
\value{
ggplot object; figure with model results
}
\description{
Takes the best linear projection* of the CATE as a function
of of user-provided covariates and visually displays resulting coefficients.
\itemize{
\item Let tau(Xi) = E(Y(1) - Y(0) | X = Xi) be the CATE, and Ai be a vector of user-provided covariates.
This function provides a (doubly robust) fit to the linear model tau(Xi) ~ beta_0 + Ai * beta."
}
}
