% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCATE.R
\name{getCATE}
\alias{getCATE}
\title{getCATE}
\usage{
getCATE(
  stdf,
  outCol,
  txCol,
  covList,
  blpredList,
  combine,
  verbose = TRUE,
  ci = 0.95,
  nTrees = 10000,
  seedN = NA
)
}
\arguments{
\item{stdf}{dataframe; trial data (single study)}

\item{outCol}{character; name of outcome column}

\item{txCol}{character; name of treatment status column}

\item{covList}{vector; column names of covariates to estimate forest}

\item{blpredList}{vector; column names of covariates to compute best linear projection; default=NULL (if NULL, same as covList)}

\item{combine}{logical; combine data from multiple studies; default=FALSE}

\item{verbose}{logical; print ATE and heterogeneity test results; default=TRUE}

\item{ci}{numeric; confidence interval \%; default=0.95}

\item{nTrees}{numeric; number of trees in the causdal forest; default=10000}

\item{seedN}{numeric; seed number to be used for reproducibility; default = NA}
}
\value{
list with:
\enumerate{
\item causal forest model; 2) ATE + CI; 3) heterogeneity test; 4) trial dataset augmented with CATE+CI;
\item best linear projection model
}
}
\description{
Takes a dataset with data coming from one or more trials and: 1) Estimates the CATE \emph{/ using a causal forest model;
2) Calculates overall ATE; 3) Carries out an agbostic test for the presence of treatment effect heterogeneity;
4) Computes the best linear projection} of CATE as a function of user-provided covariates; 6)
5) Visually displays resulting coefficients;
6) Augments original datasets with CATE + CI.

/* Let tau(Xi) = E(Y(1) - Y(0) | X = Xi) be the CATE, and Ai be a vector of user-provided covariates.
This then function provides a (doubly robust) fit to the linear model tau(Xi) ~ beta_0 + Ai * beta."
}
