% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCF.R
\name{getCF}
\alias{getCF}
\title{Takes dataframe list from multople trials and: 1) estimates CATE using causal forest;
2) Estimates overall ATE; 3) Tests for (agnostic) treatment effect heterogeneity
4) Computes the best linear projection* of CATE of user-provided covariates and visually displays resulting coefficients;
5) Augments original datasets with CATE + CI.}
\usage{
getCF(stdf, outCol, txCol, covList, blpredList, combine, ci, nTrees, seedN)
}
\arguments{
\item{stdf}{dataframe; trial data (single study)}

\item{outCol}{character; name of outcome column}

\item{txCol}{character; name of treatment status column}

\item{covList}{vector; column names of covariates to estimate forest}

\item{blpredList}{vector;}

\item{combine}{logical; combine data from multiple studies}

\item{ci}{numeric; confidence interval \%}

\item{nTrees}{numeric; column names of unit level variables  to sample units on}

\item{seedN}{numeric; seed number to be used; default = NA}
}
\value{
list with: 1) , 2),  3)
}
\description{
\itemize{
\item Let tau(Xi) = E(Y(1) - Y(0) | X = Xi) be the CATE, and Ai be a vector of user-provided covariates.
This function provides a (doubly robust) fit to the linear model tau(Xi) ~ beta_0 + Ai * beta."
}
}
