% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTauHatFig.R
\name{getTauHatFig}
\alias{getTauHatFig}
\title{Takes a dataset augmented with individual CATE estimates and visualizes overall tau.hat distribution (CATE + CI) .}
\usage{
getTauHatFig(cfdf, combine)
}
\arguments{
\item{cfdf;}{dataframe (study data augmented with cate and ci)}
}
\value{
ggplot object; figure with individual CATE estimate visualizations
}
\description{
Takes a dataset augmented with individual CATE estimates and visualizes overall tau.hat distribution (CATE + CI) .
}
