% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupCATE.R
\name{subgroupCATE}
\alias{subgroupCATE}
\title{subgroupCATE
Takes a dataset augmented with individual CATE estimates and visualizes tau.hat distribution (CATE + CI)
grouped by specified subgroups to assess HTE.}
\usage{
subgroupCATE(cfdf, outCol, txCol, covList, combine)
}
\arguments{
\item{outCol}{character; name of outcome column}

\item{txCol}{character; name of treatment status column}

\item{combine}{logical; combine data from multiple studies}

\item{cfdf;}{dataframe (study data augmented with cate and ci)}

\item{catCols}{vector; column names of categorical covariates to subgroup}

\item{numCols}{vector; column names numerical covariates to subgroup}
}
\value{
list; list of ggplot objects; figure with individual CATE estimate visualizations by subgroup
}
\description{
subgroupCATE
Takes a dataset augmented with individual CATE estimates and visualizes tau.hat distribution (CATE + CI)
grouped by specified subgroups to assess HTE.
}
