% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTauHat.R
\name{vizTauHat}
\alias{vizTauHat}
\title{vizTauHat
Takes a dataset augmented with individual CATE estimates and visualizes overall tau.hat distribution (CATE + CI) .}
\usage{
vizTauHat(cfdf, combine)
}
\arguments{
\item{cfdf;}{dataframe (study data augmented with cate and ci)}
}
\value{
ggplot object; figure with individual CATE estimate visualizations
}
\description{
vizTauHat
Takes a dataset augmented with individual CATE estimates and visualizes overall tau.hat distribution (CATE + CI) .
}
