% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCATE.R
\name{getCATE}
\alias{getCATE}
\title{getCATE}
\usage{
getCATE(stdf, outCol, txCol, covList, blpredList, combine, ci, nTrees, seedN)
}
\arguments{
\item{stdf}{dataframe; trial data (single study)}

\item{outCol}{character; name of outcome column}

\item{txCol}{character; name of treatment status column}

\item{covList}{vector; column names of covariates to estimate forest}

\item{blpredList}{vector;}

\item{combine}{logical; combine data from multiple studies}

\item{ci}{numeric; confidence interval \%}

\item{nTrees}{numeric; column names of unit level variables  to sample units on}

\item{seedN}{numeric; seed number to be used; default = NA}
}
\value{
list with: 1) , 2),  3)
}
\description{
Takes a dataset with data coming from one or more trials and: 1) Estimates the CATE \emph{/ using a causal forest model;
2) Calculates overall ATE; 3) Carries out an agbostic test for the presence of treatment effect heterogeneity;
4) Computes the best linear projection} of CATE as a function of user-provided covariates; 6)
5) Visually displays resulting coefficients;
6) Augments original datasets with CATE + CI.

/* Let tau(Xi) = E(Y(1) - Y(0) | X = Xi) be the CATE, and Ai be a vector of user-provided covariates.
This then function provides a (doubly robust) fit to the linear model tau(Xi) ~ beta_0 + Ai * beta."
}
